SdpUtils: use unifiedPlanUtils instead of plainRtpUtils
The way plainRtpUtils parses SDP to obtain encodings information (just a
single, main SSRC of the media) doesn't work for WebRTC, where there is
an SSRC group containing both a main SSRC and a secondary SSRC of the
RTX/retransmissions flow.

Instead, using Unified Plan implementation, we are able to correctly
parse the RTX SSRC and this will later be used by mediasoup to identify
and handle RTX packets.