SdpUtils: Complement encodings with both "ssrcs" and "rids"
On Chrome, an SDP Offer will come with either a list of SSRCs or a list
of RIDs, that's why the code was filling encodings from either one or
the other.

However, Firefox does indeed signal the SSRCs even when using RIDs for
simulcast, which means that both types of identifiers should be
processed.