SdpUtils: Rename sendParams, recvCaps to producerParams, consumerCaps
These names express the intention of the value with more clarity. The
method "sdpToSendRtpParameters" is really concerned with generating an
RtpParameters object that is appropriate for making a Producer out of an
SDP Offer, so "sdpToProducerRtpParameters" fits better its purpose.

This also is useful to separate our view of "sending" (a remote endpoint
sent an SDP Offer and is about to send data to us) from the view of
"sending" that is assumed by mediasoup-client (we are sending an SDP
Offer and about to start sending data).