Work around sdp-transform bug #94: "Type inconsistency in payloads"
When the SDP media line contains just 1 PayloadType number, this is
converted into an int instead of the usual string.

This has no impact in our code right now, but for consistency's sake,
it's better that small bugs like this are not allowed to exist. The
upstream library has an open issue, but doesn't look like it will get
fixed any time soon... so apply an easy workaround here.