Fixing module resolution within IDEs (#592)
* Fixing tsconfigs to help IDEs resolve imports.

Without these changes the IDE fails to find some imports and made it
hard for things like intellisense and auto complete etc.
It shouldn't make a difference with actual builds because the
moduleResolution is overridden anyway.

* Explicitly setting CJS projects to node10 moduleResolution.

node10 is the default moduleResolution for cjs projects so this is just
explicitly setting that so that the bundler entry from the base gets
properly overwritten.