Fixing a small issue with how path was setup previously. (#611)
If the node install was installed locally previously the script would
not pick that up and try to install again. It would fail because the
directory already exists and then the path would not get properly setup
for the local install of node.
This change sets PATH to point to the local install anyway. If it doesn't
exist it doesn't matter, if it does exist we use the local install.