chore: Copy package-lock to publishing packages.
Ok the package.json file needs to be deleted otherwise the packages will
try to access the local dependencies which will not be built with the
way things are setup currently. We need to remove package.json to force
npm to fetch the published packages.
By copying the package-lock file we also get the needed benefit of still
locking the required packages for the build.