chore: Updating CJS configurations of published libraries. (#620)
We were using a pretty old moduleresolution type 'node10' to build the
commonjs versions of our libraries. This was causing issues when using
some dependencies.
Updating it to 'node16' then complains that the module also needs to be
'node16'. I Previously thought that the module type needed to remain
'commonjs' to output cjs libraries but this is incorrect.
The module type of 'node16' also allows commonjs output and is more
modern so will prevent issues with other modules.