Adding .npmignore files to allow the correct things get published. (#622)
* fix: Adding npm ignore files because the gitignore files ignore types.

When there is no .npmignore file, npm uses .gitignore instead. Because
of some of the ignores in there the types end up not published. This is
a bad thing.

* chore: Bumping patch version of packages after adding npmignore.

The previous packages published will not have the types published. This
triggers a new release that should hopefully publish the correct files.

* Overrode glob to v7 for fs-routes which should solve issues with openapi paths on windows.