fix(wilbur): Fixing issues with passing arguments to wilbur. (Linux)
peer_options was always being built and passed to wilbur with the
platform_scripts. This meant that if you wished to pass your own
peer_options you could not use the platform_scripts.
This fixes that by checking that peer_options has not already been
supplied as a parameter.
Additionally, there is also now a clear delimeter between script
arguments and command arguments.