Fixing argument issues with Wilbur and platform scripts (#653)
* fix(wilbur): Fixing issues with passing arguments to wilbur. (Linux)

peer_options was always being built and passed to wilbur with the
platform_scripts. This meant that if you wished to pass your own
peer_options you could not use the platform_scripts.
This fixes that by checking that peer_options has not already been
supplied as a parameter.
Additionally, there is also now a clear delimeter between script
arguments and command arguments.

* fix(wilbur): Fixing issues with passing arguments. (Windows)

Similar to the linux changes this change should allow --peer_options and
--http_root to be passed to the server without being stomped on by the
script.

* feat(wilbur): Adding --peer_options_file to wilburs cli arguments.

This is to make it easier to provide JSON data to the peer_options
parameter without having to escape JSON data on the command line.

* chore(actions): Updating actions to match the new platform_script args.

* fix(actions): Fixed missing comma in argument list.

* fix(extras): Updating the docker commands for the new arguments.

* chore(changeset): Adding changeset for wilbur.

* Update file not found error message

Co-authored-by: Luke Bermingham <1215582+lukehb@users.noreply.github.com>

---------

Co-authored-by: Luke Bermingham <1215582+lukehb@users.noreply.github.com>